Drastically re-organised the code in the repository.
Moved all source files into locations that I felt would be more
reasonable.

Re-wrote the CMake configuration to be more friendly towards being
included by other projects, and to make it possible to fine-tune
the build configuration of the project a bit better.

Re-wrote how "externals" would be set up by the project.

Updated the CI configuration to test the code in more configurations,
and fixed the name of the "main branch".