Re-wrote the CI build configuration, and included CUDA tests in it.
In the new setup the "native" builds only test that everything (beside
SMatrix) can be turned on at the same time.

The "container" test using the non-GPU image is meant to test all of
the different build configurations. And the GPU-capable images are
meant to test everything that they can build, at once.