Removed the algebra::scalar type definition.
Instead making all of the code templated on the scalar type.
Making the user choose the scalar type that they want to use,
in "their code".

Updated the CUDA test code to produce more numerically sound
results. But even with that, had to disable all fast math in
order to achieve full agreement between the host and the device
calculations at float precision.