Made the vector(...) function "overloaded and not templated".
As it turns out, the compiler can not actually figure out the
scalar type in the expression that we use in that function for
the vc_vc frontent. So it seemed easier to just provide the
function for both float and double at the same time.

Will have to find a better way in the future if we ever want to
use some other scalar type as well.