Made all math code pick up math functions from sycl:: when using a SYCL compiler.
This is to avoid linking errors coming from unsupported standard math
functions on certain SYCL backends. Since supporting the standard functions
in device code is not part of the SYCL2020 standard itself.

Introduced a new library (algebra::common_math) to provide a central place
for the declaration of all of the standard math functions used by the project.