Refactored the CUDA test code, and introduced some SYCL tests.
In order to introduce SYCL tests, the non-CUDA-specific parts of
the CUDA test were split into a separate (algebra::tests_accel_common)
library.

The SYCL tests are defined much in the same way as the CUDA tests,
with only minor differences required by the different language.

Did not include Eigen tests, as I was unable to make use of Eigen
objects/maths in SYCL device code with the current version of Eigen.