Re-worked how compiler flags would be set up for the project.
The "externals" are no longer exposed to the flags used by this project,
so that warnings would not show up in their builds. At the same time the
include directories of Eigen and Vc were re-categorised as "system"
include directories, even when they are built as part of the project.
Thereby disabling any warnings coming from their headers.

The Debug build received -Werror (or equivalent) on all platforms now,
to make sure that warnings would be flagged by the CI system.

While that -Werror setting would not affect it, also worked around a
type of warning coming from the Vc build with GCC, which has been
bugging me since forever.