Fixed a warning coming from MSVC in algebra::cmath::vector_getter.
It showed up while trying to build detray with MSVC.

Added code for triggering the issue, so that this project would check
that function explicitly as well. Even if the return value of that
function is not checked. Because that return value is not independent
of the plugin type being used...