Introduced algebra::getter::element for each frontend.
In all cases made sure that the pre-existing element_getter functors
would be put into a single location, and all users would use that
single definition from now on.

Moved the block_getter functors into central locations as well,
although that was just for cleanup, it was not needed for the
matrix handling (yet).