Re-wrote how externals are set up by the project.
The build configuration now fully relies on the feature introduced
into FetchContent with CMake 3.24, where FetchContent will itself
choose whether to use find_package(...) or not, and what to do when
the package is found or not.

Switching out the custom cache variables used in the project for
the cache variables defined by CMake.