Restructure matrix testing code
The current matrix testing code is a bit of a mess, with e.g. 2x2 matrix
code hiding in the 3x3 matrix tests. This commit restructures the tests
to make it clearer which tests live where. Also uses macros to make
initializing the tests easier across a large range of front- and
backends.