Add SYCL CI with AMD BE.
Didn't add a test with an NVIDIA backend, as the oneAPI+CUDA combination
used by the CI crashes on the current code of the project. (The latest
versions of both do succeed however. So there's no point in debugging
this further.)

Updated (almost) all CI tests to use v67 of the Docker images.