SYCL2020 Updates, main branch (2024.11.19.) (#136)
* Update to vecmem 1.12.0.

* Made the code SYCL2020 compatible.

Using <sycl/sycl.hpp> as the main SYCL include, and picking
up all types from the ::sycl namespace.

* Update to vecmem 1.13.0.

* Add SYCL CI with AMD BE.

Didn't add a test with an NVIDIA backend, as the oneAPI+CUDA combination
used by the CI crashes on the current code of the project. (The latest
versions of both do succeed however. So there's no point in debugging
this further.)

Updated (almost) all CI tests to use v67 of the Docker images.