feat: vectorized implementation of the cross poduct (#155)
Use shuffle operations for the implementation of the cross product in the Vc AoS plugin. Since there is a trailing zero element to reach a fixed alignment, this has to be initialized to a specific value from the input vectors, so that the shuffles result in the correct vectors to do the multiplication and subtraction with.

Edit: For some reason the shuffles don't compile into something efficient on avx registers, so I kept the old fma approach for double precision