Implements an approximate equals function that works on algebra types (#160)
Implements an approximate equals function that works on algebra types. It is not foolproof for the comparison of any floating point numbers, so the google test version should be used in any actual tests.

Also adds a few missing typedefs to the algebras and fixes the mat_scalar_t type in the storage matrix and the comparison of the storage vector types, which came up during writing the tests for this PR